Fixed Issue #69 by re-implementing the tool manager. Started to add minor support for notifying tools about thread init/exit. This not finished, and therefore not dependable. Fixed an issue in the new inline assembly parser where num_immediates was not being reset. Fixed Issue #70 by removing the transparent returns tool and making it internal rather than tool-based.

+++ clients/data_collider/client.cc
@@ -201,7 +201,9 @@ class DataCollider : public MemOpInstrumentationTool {
   // therefore won't have added their shadow structure descriptions yet. We
   // need those shadow structure descriptions to determine the size of shadow
   // memory.
-  virtual void Init(InitReason) {
+  static void Init(InitReason reason) {
+    if (kInitThread == reason) return;
+
     gShiftAmount = static_cast<size_t>(
         __builtin_ctz(static_cast<unsigned>(FLAG_watchpoint_granularity)));
 
@@ -229,7 +231,8 @@ class DataCollider : public MemOpInstrumentationTool {
   }
 
   // Exit; this kills off the monitor thread.
-  virtual void Exit(ExitReason reason) {
+  static void Exit(ExitReason reason) {
+    if (kExitThread == reason) return;
     if (kExitProgram != reason && -1 != gMonitorThread) {
       kill(gMonitorThread, SIGKILL);
     }
